refactor: migrate preference configurations from preference_data to preferences module
